4_database/webapp/requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysql-connector-python==8.0.6
MySQL-python==1.2.5
urllib3==1.18
Werkzeug==0.11.11
